Fixed ATs

+++ backend/src/main/java/com/goodtimes/events/EventsRepository.java
@@ -1,11 +1,9 @@
 package com.goodtimes.events;
 
-import com.goodtimes.events.GoodtimeEvent;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-/**
- * Created by rockem on 16/2/15.
- */
-public interface EventsRepository extends MongoRepository<GoodtimeEvent, Long> {
+import java.math.BigInteger;
+
+public interface EventsRepository extends MongoRepository<GoodtimeEvent, BigInteger> {
 
 }
